+++ quasar.conf.js
@@ -123,13 +123,7 @@ module.exports = function (ctx) {
     pwa: {
       workboxOptions: {
         skipWaiting: true,
-        clientsClaim: true,
-        runtimeCaching: [
-          {
-            urlPattern: /^https:\/\/((localhost:9005)|(flespi\.io))(\/gw\/|\/auth\/|\/platform\/|\/storage\/|\/mqtt\/)/,
-            handler: 'networkOnly'
-          }
-        ]
+        clientsClaim: true
       },
       manifest: {
         name: 'Flespi ToolboX',